+++ lib/Elastica/Aggregation/AbstractAggregation.php
@@ -19,6 +19,11 @@ abstract class AbstractAggregation extends Param implements NameableInterface
     protected $_aggs = [];
 
     /**
+     * @var array|null Metadata belonging to this aggregation
+     */
+    protected $_meta;
+
+    /**
      * @param string $name the name of this aggregation
      */
     public function __construct($name)
@@ -81,6 +86,54 @@ abstract class AbstractAggregation extends Param implements NameableInterface
     }
 
     /**
+     * Add metadata to the aggregation.
+     *
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/agg-metadata.html
+     * @see \Elastica\Aggregation\AbstractAggregation::getMeta()
+     * @see \Elastica\Aggregation\AbstractAggregation::clearMeta()
+     *
+     * @param array $meta Metadata to be attached to the aggregation
+     *
+     * @return $this
+     */
+    public function setMeta(array $meta)
+    {
+        $this->_meta = $meta;
+
+        return $this;
+    }
+
+    /**
+     * Retrieve the currently configured metadata for the aggregation
+     *
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/agg-metadata.html
+     * @see \Elastica\Aggregation\AbstractAggregation::setMeta()
+     * @see \Elastica\Aggregation\AbstractAggregation::clearMeta()
+     *
+     * @return array|null
+     */
+    public function getMeta()
+    {
+        return $this->_meta;
+    }
+
+    /**
+     * Clears any previously set metadata for this aggregation.
+     *
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/agg-metadata.html
+     * @see \Elastica\Aggregation\AbstractAggregation::setMeta()
+     * @see \Elastica\Aggregation\AbstractAggregation::getMeta()
+     *
+     * @return $this
+     */
+    public function clearMeta()
+    {
+        $this->_meta = null;
+
+        return $this;
+    }
+
+    /**
      * @return array
      */
     public function toArray()
@@ -90,6 +143,9 @@ abstract class AbstractAggregation extends Param implements NameableInterface
         if (array_key_exists('global_aggregation', $array)) {
             // compensate for class name GlobalAggregation
             $array = ['global' => new \stdClass()];
+        }
+        if (isset($this->_meta) && sizeof($this->_meta)) {
+            $array['meta'] = $this->_convertArrayable($this->_meta);
         }
         if (sizeof($this->_aggs)) {
             $array['aggs'] = $this->_convertArrayable($this->_aggs);